Began GUI
Wrote a system that starts the controller thread, with an LLController object, in a  separate process using the multiprocessing module.  Data is shared between the main gui process and this controller process via named memory mapped files (using mmap), synchronized using NamedMutexes.  A single standard inherited queue is used to create LLControllerInterface connections from other processes to the main  LLController interface,and pipes which are sent over these queues when an LLControllerInterface is connecting are then used  to manage each connection.